Refactor uses with the new value objects

+++ app/models/todo/item/delete.rb
@@ -2,19 +2,19 @@
 
 module Todo::Item
   class Delete < ::Micro::Case
-    attribute :id, validates: {numericality: {only_integer: true}}
-    attribute :user_id, validates: {numericality: {only_integer: true}}
+    attribute :id, default: proc(&::Kind::ID)
+    attribute :user_id, default: proc(&::Kind::ID)
     attribute :repository, {
       default: Repository,
       validates: {kind: {respond_to: :delete_item}}
     }
 
     def call!
+      return Failure(:invalid_scope) if id.invalid? || user_id.invalid?
+
       deleted = repository.delete_item(user_id:, id:)
 
-      return Failure(:todo_not_found) unless deleted
-
-      Success(:todo_deleted)
+      deleted ? Success(:todo_deleted) : Failure(:todo_not_found)
     end
   end
 end
